requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
google-generativeai==0.7.2
requests==2.32.3
streamlit==1.38.0
pandas==2.2.3
